backend/requirements.txt
django>=4.2,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<5
httpx>=0.23.0,<0.25.0
python-dotenv>=1.0,<2
trimesh>=4,<5
numpy>=1.26,<2
requests>=2.31,<3
dj-database-url>=2.1,<3
gunicorn>=21.2,<22
requests-oauthlib>=1.3,<2
supabase>=1.0.0,<2
google-genai>=1.5.0,<2